engine_traits_tests: Move mods into files
